type: feature
subject: added a basic json file for configuration. so far, it contains
rules regarding whether the output should contain errors and whether a
rule is enabled or not. also added a command for configuration which
empty so far. the plan is to implement getter and most importantly
setter subcommands that can manipulate the config file.