type: feature
subject: implemented a "check" command that can check a file for errors.
prints all errors found by default (configurable). this might be a
placeholder feature depending on feedback from Michael. if he likes it,
an idea is to allow multiple file names as arguments in order to check
multiple files at a time