type: feature
subject: implemented a print command that prints an input file to the
terminal. can be configured to lint as well as print errors found. this
might be a placeholder command as it has no real world use except for
demonstrating that the logic of the application works in an easy way