type: structure
subject: the data folder is currently the path that the program looks
for files in. it also contains the /output/-folder which is where the
write command writes its output file to. depending on how development
goes this might be deprecated to leave room for relative paths in the
program. otherwise it would make sense to make the path configurable