type: package
subject: lalalint is the new /pkg/, it currently contains all the main
logic of the program. the current state of the directory is not up to
par and needs to be addressed. smaller files or possibly sub-packages
would make sense to divide the program into smaller pieces with a more
clear purpose.